Rename framework to BLTNBoard
