+++ Example/Swift/Bulletin/BulletinDataSource.swift
@@ -4,7 +4,7 @@
  */
 
 import UIKit
-import Bulletin
+import BLTNBoard
 import SafariServices
 
 /**
